+++ driverlib/MSP430FR5xx_6xx/eusci_a_uart.h
@@ -1,34 +1,3 @@
-/* --COPYRIGHT--,BSD
- * Copyright (c) 2016, Texas Instruments Incorporated
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * *  Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- *
- * *  Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * *  Neither the name of Texas Instruments Incorporated nor the names of
- *    its contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- * --/COPYRIGHT--*/
 //*****************************************************************************
 //
 // eusci_a_uart.h - Driver for the EUSCI_A_UART Module.
@@ -67,8 +36,7 @@ extern "C"
 //! \brief Used in the EUSCI_A_UART_init() function as the param parameter.
 //
 //*****************************************************************************
-typedef struct EUSCI_A_UART_initParam
-{
+typedef struct EUSCI_A_UART_initParam {
     //! Selects Clock source. Refer to device specific datasheet for available
     //! options.
     //! \n Valid values are:
@@ -114,6 +82,7 @@ typedef struct EUSCI_A_UART_initParam
     //! - \b EUSCI_A_UART_LOW_FREQUENCY_BAUDRATE_GENERATION
     uint8_t overSampling;
 } EUSCI_A_UART_initParam;
+
 
 //*****************************************************************************
 //
@@ -409,7 +378,7 @@ extern uint8_t EUSCI_A_UART_getInterruptStatus(uint16_t baseAddress,
 //
 //*****************************************************************************
 extern void EUSCI_A_UART_clearInterrupt(uint16_t baseAddress,
-                                        uint8_t mask);
+                                        uint16_t mask);
 
 //*****************************************************************************
 //
