+++ driverlib/MSP430FR5xx_6xx/framctl.h
@@ -1,34 +1,3 @@
-/* --COPYRIGHT--,BSD
- * Copyright (c) 2016, Texas Instruments Incorporated
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * *  Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- *
- * *  Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * *  Neither the name of Texas Instruments Incorporated nor the names of
- *    its contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- * --/COPYRIGHT--*/
 //*****************************************************************************
 //
 // framctl.h - Driver for the FRAMCTL Module.
@@ -187,7 +156,7 @@ extern void FRAMCtl_fillMemory32(uint32_t value,
 //! \return None
 //
 //*****************************************************************************
-extern void FRAMCtl_enableInterrupt(uint8_t interruptMask);
+extern void FRAMCtl_enableInterrupt(uint16_t interruptMask);
 
 //*****************************************************************************
 //
@@ -207,13 +176,13 @@ extern void FRAMCtl_enableInterrupt(uint8_t interruptMask);
 //!           FRAMCtl memory error detection logic.
 //!
 //! \return Logical OR of any of the following:
-//!         - \b FRAMCtl_ACCESS_TIME_ERROR_FLAG Interrupt flag is set if a
+//!         - \b FRAMCTL_ACCESS_TIME_ERROR_FLAG Interrupt flag is set if a
 //!         wrong setting for NPRECHG and NACCESS is set and FRAMCtl access
 //!         time is not hold.
-//!         - \b FRAMCtl_UNCORRECTABLE_BIT_FLAG Interrupt flag is set if an
+//!         - \b FRAMCTL_UNCORRECTABLE_BIT_FLAG Interrupt flag is set if an
 //!         uncorrectable bit error has been detected in the FRAMCtl memory
 //!         error detection logic.
-//!         - \b FRAMCtl_CORRECTABLE_BIT_FLAG Interrupt flag is set if a
+//!         - \b FRAMCTL_CORRECTABLE_BIT_FLAG Interrupt flag is set if a
 //!         correctable bit error has been detected and corrected in the
 //!         FRAMCtl memory error detection logic.
 //!         \n indicating the status of the masked flags
